refactored lib to use ReadWriteListener instead of BleDevice.ReadWriteListener, removed old interface
