+++ app/src/main/java/com/idevicesinc/sweetblue/PA_Task_Transactionable.java
@@ -1,12 +1,5 @@
 package com.idevicesinc.sweetblue;
 
-import java.util.UUID;
-
-import android.bluetooth.BluetoothGatt;
-
-import com.idevicesinc.sweetblue.BleDevice.ReadWriteListener.ReadWriteEvent;
-import com.idevicesinc.sweetblue.BleDevice.ReadWriteListener.Status;
-import com.idevicesinc.sweetblue.BleDevice.ReadWriteListener.Target;
 
 abstract class PA_Task_Transactionable extends PA_Task_RequiresConnection
 {
