+++ app/src/main/java/com/idevicesinc/sweetblue/PI_EndListener.java
@@ -1,6 +1,5 @@
 package com.idevicesinc.sweetblue;
 
-import com.idevicesinc.sweetblue.BleDevice.ReadWriteListener;
 import com.idevicesinc.sweetblue.BleTransaction.EndReason;
 
 interface PI_EndListener
