+++ app/src/main/java/com/idevicesinc/sweetblue/P_Task_ExecuteReliableWrite.java
@@ -7,9 +7,9 @@ import com.idevicesinc.sweetblue.utils.Utils;
 final class P_Task_ExecuteReliableWrite extends PA_Task_RequiresConnection implements PA_Task.I_StateListener
 {
 	private final PE_TaskPriority m_priority;
-	private final BleDevice.ReadWriteListener m_listener;
+	private final ReadWriteListener m_listener;
 
-	public P_Task_ExecuteReliableWrite(BleDevice device, BleDevice.ReadWriteListener listener, PE_TaskPriority priority)
+	public P_Task_ExecuteReliableWrite(BleDevice device, ReadWriteListener listener, PE_TaskPriority priority)
 	{
 		super(device, null);
 
@@ -17,14 +17,14 @@ final class P_Task_ExecuteReliableWrite extends PA_Task_RequiresConnection imple
 		m_listener = listener;
 	}
 
-	private BleDevice.ReadWriteListener.ReadWriteEvent newEvent(final BleDevice.ReadWriteListener.Status status, final int gattStatus, final boolean solicited)
+	private ReadWriteListener.ReadWriteEvent newEvent(final ReadWriteListener.Status status, final int gattStatus, final boolean solicited)
 	{
 		return getDevice().m_reliableWriteMngr.newEvent(status, gattStatus, solicited);
 	}
 
-	private void invokeListeners(final BleDevice.ReadWriteListener.Status status, final int gattStatus)
+	private void invokeListeners(final ReadWriteListener.Status status, final int gattStatus)
 	{
-		final BleDevice.ReadWriteListener.ReadWriteEvent event = newEvent(BleDevice.ReadWriteListener.Status.NOT_CONNECTED, gattStatus, /*solicited=*/true);
+		final ReadWriteListener.ReadWriteEvent event = newEvent(ReadWriteListener.Status.NOT_CONNECTED, gattStatus, /*solicited=*/true);
 
 		getDevice().invokeReadWriteCallback(m_listener, event);
 	}
@@ -33,7 +33,7 @@ final class P_Task_ExecuteReliableWrite extends PA_Task_RequiresConnection imple
 	{
 		super.onNotExecutable();
 
-		invokeListeners(BleDevice.ReadWriteListener.Status.NOT_CONNECTED, BleStatuses.GATT_STATUS_NOT_APPLICABLE);
+		invokeListeners(ReadWriteListener.Status.NOT_CONNECTED, BleStatuses.GATT_STATUS_NOT_APPLICABLE);
 	}
 
 	@Override protected BleTask getTaskType()
@@ -45,7 +45,7 @@ final class P_Task_ExecuteReliableWrite extends PA_Task_RequiresConnection imple
 	{
 		if( false == getDevice().layerManager().executeReliableWrite() )
 		{
-			fail(BleDevice.ReadWriteListener.Status.FAILED_TO_SEND_OUT, BleStatuses.GATT_STATUS_NOT_APPLICABLE);
+			fail(ReadWriteListener.Status.FAILED_TO_SEND_OUT, BleStatuses.GATT_STATUS_NOT_APPLICABLE);
 		}
 		else
 		{
@@ -59,19 +59,19 @@ final class P_Task_ExecuteReliableWrite extends PA_Task_RequiresConnection imple
 		{
 			succeed();
 
-			invokeListeners(BleDevice.ReadWriteListener.Status.SUCCESS, gattStatus);
+			invokeListeners(ReadWriteListener.Status.SUCCESS, gattStatus);
 		}
 		else
 		{
-			fail(BleDevice.ReadWriteListener.Status.REMOTE_GATT_FAILURE, gattStatus);
+			fail(ReadWriteListener.Status.REMOTE_GATT_FAILURE, gattStatus);
 		}
 	}
 
-	private void fail(final BleDevice.ReadWriteListener.Status status, final int gattStatus)
+	private void fail(final ReadWriteListener.Status status, final int gattStatus)
 	{
 		super.fail();
 
-		invokeListeners(BleDevice.ReadWriteListener.Status.REMOTE_GATT_FAILURE, gattStatus);
+		invokeListeners(ReadWriteListener.Status.REMOTE_GATT_FAILURE, gattStatus);
 	}
 
 	@Override public PE_TaskPriority getPriority()
@@ -83,7 +83,7 @@ final class P_Task_ExecuteReliableWrite extends PA_Task_RequiresConnection imple
 	{
 		if( state == PE_TaskState.TIMED_OUT )
 		{
-			getDevice().invokeReadWriteCallback(m_listener, newEvent(BleDevice.ReadWriteListener.Status.TIMED_OUT, BleStatuses.GATT_STATUS_NOT_APPLICABLE, /*solicited=*/true));
+			getDevice().invokeReadWriteCallback(m_listener, newEvent(ReadWriteListener.Status.TIMED_OUT, BleStatuses.GATT_STATUS_NOT_APPLICABLE, /*solicited=*/true));
 		}
 		else if( state == PE_TaskState.SOFTLY_CANCELLED )
 		{
