+++ app/src/main/java/com/idevicesinc/sweetblue/P_Task_ReadRssi.java
@@ -2,9 +2,9 @@ package com.idevicesinc.sweetblue;
 
 import android.bluetooth.BluetoothGatt;
 
-import com.idevicesinc.sweetblue.BleDevice.ReadWriteListener.ReadWriteEvent;
-import com.idevicesinc.sweetblue.BleDevice.ReadWriteListener.Status;
-import com.idevicesinc.sweetblue.BleDevice.ReadWriteListener.Type;
+import com.idevicesinc.sweetblue.ReadWriteListener.ReadWriteEvent;
+import com.idevicesinc.sweetblue.ReadWriteListener.Status;
+import com.idevicesinc.sweetblue.ReadWriteListener.Type;
 import com.idevicesinc.sweetblue.utils.Utils;
 
 /**
@@ -13,10 +13,10 @@ import com.idevicesinc.sweetblue.utils.Utils;
  */
 final class P_Task_ReadRssi extends PA_Task_Transactionable implements PA_Task.I_StateListener
 {	
-	protected final BleDevice.ReadWriteListener m_readWriteListener;
+	protected final ReadWriteListener m_readWriteListener;
 	private final Type m_type;
 	
-	public P_Task_ReadRssi(BleDevice device, BleDevice.ReadWriteListener readWriteListener, BleTransaction txn_nullable, PE_TaskPriority priority, Type type)
+	public P_Task_ReadRssi(BleDevice device, ReadWriteListener readWriteListener, BleTransaction txn_nullable, PE_TaskPriority priority, Type type)
 	{
 		super(device, txn_nullable, false, priority);
