+++ tester/src/test/java/com/idevicesinc/sweetblue/DuplicateCharTest.java
@@ -63,7 +63,7 @@ public class DuplicateCharTest extends BaseBleUnitTest
                                     {
                                         return mTestDesc;
                                     }
-                                }, new BleDevice.ReadWriteListener()
+                                }, new ReadWriteListener()
                                 {
                                     @Override public void onEvent(ReadWriteEvent e)
                                     {
@@ -115,7 +115,7 @@ public class DuplicateCharTest extends BaseBleUnitTest
                                     {
                                         return mTestDesc;
                                     }
-                                }, new BleDevice.ReadWriteListener()
+                                }, new ReadWriteListener()
                                 {
                                     @Override public void onEvent(ReadWriteEvent e)
                                     {
@@ -169,7 +169,7 @@ public class DuplicateCharTest extends BaseBleUnitTest
                                     {
                                         return mTestDesc;
                                     }
-                                }, new BleDevice.ReadWriteListener()
+                                }, new ReadWriteListener()
                                 {
                                     @Override public void onEvent(ReadWriteEvent e)
                                     {
